+++ apps/vmq_server/test/vmq_rate_limiter_SUITE.erl
@@ -1,17 +1,7 @@
 -module(vmq_rate_limiter_SUITE).
--export([
-         %% suite/0,
-         init_per_suite/1,
-         end_per_suite/1,
-         init_per_testcase/2,
-         end_per_testcase/2,
-         all/0
-        ]).
 
--export([publish_rate_limit_test/1]).
-
--export([hook_auth_on_register/5,
-         hook_auth_on_publish/6]).
+-compile(export_all).
+-compile(nowarn_export_all).
 
 %% ===================================================================
 %% common_test callbacks
@@ -19,43 +9,66 @@
 init_per_suite(Config) ->
     S = vmq_test_utils:get_suite_rand_seed(),
     cover:start(),
-    [S|Config].
+    vmq_test_utils:setup(),
+    vmq_server_cmd:set_config(allow_anonymous, false),
+    vmq_server_cmd:set_config(retry_interval, 10),
+    vmq_server_cmd:set_config(max_client_id_size, 100),
+    vmq_server_cmd:listener_start(1888, [{allowed_protocol_versions, "3,4,5"}]),
+    [S,{ct_hooks, vmq_cth}| Config].
 
 end_per_suite(_Config) ->
+    vmq_test_utils:teardown(),
     _Config.
+
+init_per_group(mqttv4, Config) ->
+    [{protover, 4}|Config];
+init_per_group(mqttv5, Config) ->
+    [{protover, 5}|Config].
+
+end_per_group(_Group, _Config) ->
+    ok.
 
 init_per_testcase(_Case, Config) ->
     vmq_test_utils:seed_rand(Config),
-    vmq_test_utils:setup(),
-    vmq_server_cmd:set_config(retry_interval, 10),
-    vmq_server_cmd:set_config(allow_anonymous, false),
-    vmq_server_cmd:listener_start(1888, []),
     Config.
 
 end_per_testcase(_, Config) ->
-    vmq_test_utils:teardown(),
     Config.
 
 all() ->
-    [publish_rate_limit_test].
+    [{group, mqttv4},
+     {group, mqttv5}].
+
+groups() ->
+    Tests = [publish_rate_limit_test],
+    [
+     {mqttv4, [], Tests},
+     {mqttv5, [], Tests}
+    ].
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%% Actual Tests
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-publish_rate_limit_test(_) ->
+publish_rate_limit_test(Cfg) ->
     %% Rate Limit is enfored in auth_on_register hook
-    Connect = packet:gen_connect("rate-limit-test", [{keepalive, 60}]),
-    Connack = packet:gen_connack(0),
+    ClientId = vmq_cth:ustr(Cfg),
+    Username = <<"rate-limit-test">>,
+    Password = <<"secret">>,
+
+    Connect = mqtt5_v4compat:gen_connect(ClientId, [{username, Username},
+                                                    {password, Password}], Cfg),
+    Connack = mqtt5_v4compat:gen_connack(success, Cfg),
     Pub = fun(Sleep, Socket, Id) ->
-                  Publish = packet:gen_publish("rate/limit/test", 1,
-                                               vmq_test_utils:rand_bytes(1460), [{mid, Id}]),
-                  Puback = packet:gen_puback(Id),
+                  Payload = vmq_test_utils:rand_bytes(1460),
+                  Publish = mqtt5_v4compat:gen_publish("rate/limit/test", 1,
+                                                       Payload, [{mid, Id}], Cfg),
+                  Puback = mqtt5_v4compat:gen_puback(Id, Cfg),
                   ok = gen_tcp:send(Socket, Publish),
-                  ok = packet:expect_packet(Socket, "puback", Puback),
+                  ok = mqtt5_v4compat:expect_packet(Socket, "puback", Puback, Cfg),
                   timer:sleep(Sleep)
           end,
     enable_hooks(),
-    {ok, Socket} = packet:do_client_connect(Connect, Connack, []),
+    {ok, Socket} = mqtt5_v4compat:do_client_connect(Connect, Connack, [], Cfg),
     {T, _} =
     timer:tc(
       fun() ->
@@ -77,11 +90,19 @@ publish_rate_limit_test(_) ->
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%% Hooks (as explicit as possible)
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-hook_auth_on_register(_Peer, {"", <<"rate-limit-test">>}, _User, _Password, _Clean) ->
+hook_auth_on_register(_Peer, {"", _}, <<"rate-limit-test">>, _Password, _Clean) ->
     %% this will limit the publisher to 1 message/sec
     {ok, [{max_message_rate, 1}]}.
 
+hook_auth_on_register_m5(_Peer, {"", _}, <<"rate-limit-test">>, _Password, _CleanStart, _) ->
+    %% this will limit the publisher to 1 message/sec
+    {ok, #{max_message_rate => 1}}.
+
 hook_auth_on_publish(_, _, _MsgId, _, _, _) -> ok.
+hook_auth_on_publish_m5(_, _, _MsgId, _, _, _, _) -> ok.
+
+
+
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%% Helper
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -89,9 +110,22 @@ enable_hooks() ->
     vmq_plugin_mgr:enable_module_plugin(
       auth_on_register, ?MODULE, hook_auth_on_register, 5),
     vmq_plugin_mgr:enable_module_plugin(
-      auth_on_publish, ?MODULE, hook_auth_on_publish, 6).
+      auth_on_publish, ?MODULE, hook_auth_on_publish, 6),
+
+    vmq_plugin_mgr:enable_module_plugin(
+      auth_on_register_m5, ?MODULE, hook_auth_on_register_m5, 6),
+    vmq_plugin_mgr:enable_module_plugin(
+      auth_on_publish_m5, ?MODULE, hook_auth_on_publish_m5, 7).
+
+
+
 disable_hooks() ->
     vmq_plugin_mgr:disable_module_plugin(
       auth_on_register, ?MODULE, hook_auth_on_register, 5),
     vmq_plugin_mgr:disable_module_plugin(
-      auth_on_publish, ?MODULE, hook_auth_on_publish, 6).
+      auth_on_publish, ?MODULE, hook_auth_on_publish, 6),
+
+    vmq_plugin_mgr:disable_module_plugin(
+      auth_on_register_m5, ?MODULE, hook_auth_on_register_m5, 6),
+    vmq_plugin_mgr:disable_module_plugin(
+      auth_on_publish_m5, ?MODULE, hook_auth_on_publish_m5, 7).