+++ changelog.md
@@ -5,6 +5,7 @@
 - Fix `vmq_webhooks` issue where MQTTv5 hooks where not configurable in the
   `vernemq.conf` file.
 - Fix bug in Prometheus output (#923).
+- Fix `max_message_rate` to include MQTTv5 sessions.
 
 ## VerneMQ 1.6.0